requirements.txt
beautifulsoup4==4.9.1
blinker==1.8.2
cachecontrol==0.14.2
cachetools==5.5.2
certifi==2020.6.20
cffi==1.17.1
chardet==3.0.4
click==8.1.8
cryptography==44.0.2
cycler==0.10.0
firebase-admin==6.7.0
flask==3.0.3
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==19.9.0
httplib2==0.22.0
idna==2.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.3.1
MarkupSafe==2.1.5
matplotlib==3.3.2
msgpack==1.1.0
numpy==1.21.0
Pillow==9.0.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2025.2
requests==2.24.0
rsa==4.9.1
six==1.15.0
soupsieve==2.0.1
sqlalchemy==2.0.40
typing-extensions==4.13.2
uritemplate==4.1.1
urllib3==1.25.11
werkzeug==3.0.6
zipp==3.20.2
